requirements.txt
streamlit==1.39.0
numpy==1.26.4
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
pillow==10.4.0
scikit-learn==1.4.2
